Added checking for enough rooms in _assign_competing_teams
To make code cleaner, the functionality for _assign_competing_teams was
split into two functions, one with the original name which only cares
about categorising a list of attendances to compete or judge, and a new
function assign_teams_today() which gets the attendances for the day and
generates the corresponding MatchDay if _assign_competing_teams gave
a valid result. The advantage of this is that the room checking logic
resides in the assign_teams_today() function which is more natural.